backend/requirements.txt
pyupbit==0.2.33
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
aiohttp==3.9.1
pydantic==2.5.2
python-jose==3.3.0
schedule==1.2.1
requests==2.31.0



